feat: consolidate navigation into route-backed sections
