+++ app/components/themes/BuiltinThemeHeader.vue
@@ -142,12 +142,12 @@
                 </template>
                 <v-list density="compact">
                     <v-list-item
-                        to="/user/detail"
+                        to="/me/account"
                         :title="t('messages.userCenter')"
                         prepend-icon="mdi-account-box"
                     />
                     <v-list-item
-                        to="/user/history"
+                        to="/me/history"
                         :title="t('messages.readingHistory')"
                         prepend-icon="mdi-history"
                     />
@@ -206,7 +206,10 @@
             mobile-breakpoint="md"
             :width="drawerWidth"
         >
-            <v-list density="compact">
+            <v-list
+                class="tb-theme-drawer__list"
+                density="compact"
+            >
                 <template
                     v-for="item in navItems"
                     :key="item.key"
@@ -269,6 +272,7 @@
                     </v-list-item>
                 </template>
             </v-list>
+            <SidebarHelpMenu />
         </v-navigation-drawer>
     </div>
 </template>
@@ -277,6 +281,8 @@
 import { computed, onMounted, ref } from 'vue';
 import { useDisplay } from 'vuetify';
 import { useMainStore } from '@/stores/main';
+import { usePrimaryNavigation } from '@/composables/usePrimaryNavigation';
+import SidebarHelpMenu from '@/components/SidebarHelpMenu.vue';
 import { useI18n } from '#i18n';
 
 const props = defineProps({
@@ -296,7 +302,6 @@ const sidebar = ref(true);
 const mobileSearch = ref(false);
 const search = ref('');
 const searchCategory = ref('all');
-const readDoneCount = ref(0);
 
 const isLightGray = computed(() => props.variant === 'light-gray');
 const isMinimal = computed(() => props.variant === 'minimal');
@@ -326,83 +331,7 @@ const brandMark = computed(() => {
     return '书';
 });
 
-const navItems = computed(() => {
-    const items = [
-        { key: 'home', icon: 'mdi-home', href: '/', text: t('navigation.home') },
-        { key: 'library', icon: 'mdi-book', href: '/library', text: t('navigation.localLibrary') },
-    ];
-    if (store.sys.show_network_library !== false) {
-        items.push({ key: 'network', icon: 'mdi-cloud-search', href: '/network', text: t('navigation.networkLibrary') });
-    }
-    items.push({ key: 'audios', icon: 'mdi-book-music', href: '/audios', text: t('navigation.audiobooks'), badge: t('audiobook.beta') });
-    if (store.user.is_login) {
-        items.push({ key: 'shelf', icon: 'mdi-bookshelf', href: '/user/shelf', text: t('navigation.myShelf') });
-    }
-    if (store.user.is_admin) {
-        items.push({
-            key: 'admin',
-            icon: 'mdi-cog',
-            text: t('navigation.admin'),
-            groups: [
-                { icon: 'mdi-cog', href: '/admin/settings', text: t('navigation.settings') },
-                { icon: 'mdi-human-greeting', href: '/admin/users', text: t('navigation.users') },
-                { icon: 'mdi-library-shelves', href: '/admin/books', text: t('navigation.books') },
-                { icon: 'mdi-playlist-music', href: '/audio-jobs', text: t('navigation.audiobookJobs') },
-                { icon: 'mdi-import', href: '/admin/imports', text: t('navigation.import') },
-                { icon: 'mdi-power-plug-outline', href: '/admin/plugins', text: t('navigation.plugins') },
-                { icon: 'mdi-palette', href: '/admin/themes', text: t('navigation.themes') },
-                { icon: 'mdi-text-box-outline', href: '/admin/logs', text: t('navigation.systemLogs') },
-            ],
-        });
-    }
-    items.push(
-        { key: 'categories', heading: t('navigation.categories') },
-    );
-    if (store.user.is_login) {
-        items.push({
-            key: 'read-books',
-            icon: 'mdi-check-circle',
-            href: '/user/history?tab=finished',
-            text: t('navigation.readBooks'),
-            count: readDoneCount.value,
-        });
-    }
-    items.push(
-        { key: 'nav', icon: 'mdi-widgets', href: '/nav', text: t('navigation.browse'), count: store.sys.books },
-        { key: 'publisher', icon: 'mdi-home-group', href: '/publisher', text: t('navigation.publishers'), count: store.sys.publishers },
-        { key: 'author', icon: 'mdi-human-greeting', href: '/author', text: t('navigation.authors'), count: store.sys.authors },
-        { key: 'tag', icon: 'mdi-tag-heart', href: '/tag', text: t('navigation.tags'), count: store.sys.tags },
-        { key: 'format', icon: 'mdi-file', href: '/format', text: t('navigation.formats'), count: store.sys.formats },
-        { key: 'series', icon: 'mdi-library-shelves', href: '/series', text: t('navigation.series'), count: store.sys.series },
-        { key: 'rating', icon: 'mdi-star-half', href: '/rating', text: t('navigation.ratings') },
-        { key: 'hot', icon: 'mdi-trending-up', href: '/hot', text: t('navigation.hot') },
-        { key: 'recent', icon: 'mdi-history', href: '/recent', text: t('navigation.recent') },
-    );
-    if (store.sys.friends?.length > 0) {
-        items.push(
-            { key: 'friends', heading: t('messages.friendshipLinks') },
-            {
-                key: 'friend-links',
-                target: '_blank',
-                links: store.sys.friends.map(friend => ({
-                    icon: 'mdi-open-in-new',
-                    href: friend.href,
-                    text: friend.text,
-                })),
-            },
-        );
-    }
-    if (store.sys.show_sidebar_sys !== false) {
-        items.push(
-            { key: 'system', heading: t('messages.system') },
-            { key: 'version', icon: 'mdi-history', href: '', text: t('messages.systemVersion'), count: store.sys.version },
-            { key: 'users', icon: 'mdi-human', href: '', text: t('messages.userCount'), count: store.sys.users },
-            { key: 'opds', icon: 'mdi-cellphone', href: '/opds-readme', text: t('messages.opdsIntroduction'), count: 'OPDS' },
-            { key: 'webdav', icon: 'mdi-cloud-sync', href: '/webdav-readme', text: t('messages.webdavIntroduction'), count: 'WebDAV' },
-        );
-    }
-    return items;
-});
+const navItems = usePrimaryNavigation(store, t);
 
 function toggleDrawer() {
     sidebar.value = !sidebar.value;
@@ -419,14 +348,6 @@ function doSearch() {
 
 onMounted(() => {
     sidebar.value = display.mdAndUp.value;
-    if (store.user.is_login) {
-        const { $backend } = useNuxtApp();
-        $backend('/read-done').then((rsp) => {
-            if (rsp.err === 'ok') {
-                readDoneCount.value = rsp.total || 0;
-            }
-        }).catch(() => {});
-    }
 });
 
 </script>
@@ -441,6 +362,16 @@ onMounted(() => {
     gap: 8px;
     min-width: 0;
     padding: 0 10px 0 2px;
+}
+
+.tb-theme-drawer :deep(.v-navigation-drawer__content) {
+    display: flex;
+    flex-direction: column;
+}
+
+.tb-theme-drawer__list {
+    flex: 1 1 auto;
+    overflow-y: auto;
 }
 
 .tb-theme-brand-mark {
