+++ app/pages/admin/settings/general.vue
@@ -1,0 +1,7 @@
+<template>
+    <GeneralSettingsContent />
+</template>
+
+<script setup>
+import GeneralSettingsContent from '@/components/admin/GeneralSettingsContent.vue';
+</script>
