+++ app/pages/admin/settings/index.vue
@@ -1,0 +1,7 @@
+<template>
+    <div />
+</template>
+
+<script setup>
+await navigateTo('/admin/settings/general', { replace: true });
+</script>
