+++ app/pages/admin/settings/plugins.vue
@@ -1,0 +1,7 @@
+<template>
+    <PluginCenterContent />
+</template>
+
+<script setup>
+import PluginCenterContent from '@/pages/admin/plugins/index.vue';
+</script>
