+++ app/pages/admin/settings/themes.vue
@@ -1,0 +1,7 @@
+<template>
+    <ThemeManagementContent />
+</template>
+
+<script setup>
+import ThemeManagementContent from '@/pages/admin/themes.vue';
+</script>
