+++ app/pages/library/index.vue
@@ -1,0 +1,7 @@
+<template>
+    <div />
+</template>
+
+<script setup>
+await navigateTo('/library/local', { replace: true });
+</script>
