+++ app/pages/library/local.vue
@@ -1,0 +1,7 @@
+<template>
+    <LocalLibraryContent />
+</template>
+
+<script setup>
+import LocalLibraryContent from '@/components/library/LocalLibraryContent.vue';
+</script>
