+++ app/pages/library/network.vue
@@ -1,0 +1,7 @@
+<template>
+    <NetworkLibraryContent />
+</template>
+
+<script setup>
+import NetworkLibraryContent from '@/pages/network/index.vue';
+</script>
