+++ app/pages/me/account.vue
@@ -1,0 +1,7 @@
+<template>
+    <AccountContent />
+</template>
+
+<script setup>
+import AccountContent from '@/pages/user/detail.vue';
+</script>
