+++ app/pages/me/history.vue
@@ -1,0 +1,7 @@
+<template>
+    <HistoryContent />
+</template>
+
+<script setup>
+import HistoryContent from '@/pages/user/history.vue';
+</script>
