+++ app/pages/me/index.vue
@@ -1,0 +1,7 @@
+<template>
+    <div />
+</template>
+
+<script setup>
+await navigateTo('/me/shelf', { replace: true });
+</script>
