+++ app/pages/me/private.vue
@@ -1,0 +1,7 @@
+<template>
+    <PrivateBooksContent />
+</template>
+
+<script setup>
+import PrivateBooksContent from '@/pages/scopedbooks.vue';
+</script>
