+++ app/pages/me/shelf.vue
@@ -1,0 +1,7 @@
+<template>
+    <ShelfContent />
+</template>
+
+<script setup>
+import ShelfContent from '@/pages/user/shelf.vue';
+</script>
