+++ app/test/components/AppHeader.spec.ts
@@ -6,7 +6,7 @@ import * as components from 'vuetify/components';
 import * as directives from 'vuetify/directives';
 import { beforeEach, describe, expect, it, vi } from 'vitest';
 
-vi.mock('vue-i18n', () => ({
+vi.mock('#i18n', () => ({
     useI18n: () => ({
         locale: { value: 'zh-CN' },
         locales: { value: [{ code: 'zh-CN', name: '简体中文' }] },
@@ -65,6 +65,8 @@ describe('AppHeader.vue', () => {
     beforeEach(() => {
         pushMock.mockReset();
         storeState.sys.show_network_library = true;
+        storeState.user.is_login = false;
+        storeState.user.is_admin = false;
     });
 
     it('only wraps the site title text in the clickable/pointer area, not the whole title bar', () => {
@@ -92,19 +94,37 @@ describe('AppHeader.vue', () => {
         wrapper.unmount();
     });
 
-    it('shows the network library link by default', () => {
+    it('shows one consolidated library entry regardless of the legacy visibility setting', () => {
         const wrapper = mountHeader();
 
-        expect(wrapper.text()).toContain('navigation.networkLibrary');
+        expect(wrapper.text()).toContain('navigation.libraryBrowse');
+        expect(wrapper.text()).not.toContain('navigation.localLibrary');
+        expect(wrapper.text()).not.toContain('navigation.networkLibrary');
+        storeState.sys.show_network_library = false;
+        expect(wrapper.text()).toContain('navigation.libraryBrowse');
 
         wrapper.unmount();
     });
 
-    it('hides the network library link when disabled', () => {
-        storeState.sys.show_network_library = false;
+    it('puts My Reading immediately after Home for signed-in users', () => {
+        storeState.user.is_login = true;
         const wrapper = mountHeader();
+        const labels = wrapper.findAll('.v-list-item-title').map(item => item.text());
 
-        expect(wrapper.text()).not.toContain('navigation.networkLibrary');
+        expect(labels.indexOf('navigation.myReading')).toBe(labels.indexOf('navigation.home') + 1);
+        expect(labels).not.toContain('navigation.readBooks');
+
+        wrapper.unmount();
+    });
+
+    it('keeps plugin and theme management inside System Settings', () => {
+        storeState.user.is_admin = true;
+        const wrapper = mountHeader();
+        const links = wrapper.findAllComponents({ name: 'VListItem' }).map(item => item.props('to'));
+
+        expect(links).toContain('/admin/settings/general');
+        expect(links).not.toContain('/admin/plugins');
+        expect(links).not.toContain('/admin/themes');
 
         wrapper.unmount();
     });
