+++ app/test/e2e/settings.spec.ts
@@ -10,19 +10,20 @@ test.describe('Admin Settings (GitHub-style layout)', () => {
         });
     });
 
-    test('顶部悬浮标题栏含标题与保存按钮', async ({ page }) => {
+    test('路径页签与保存栏在滚动时保持可见', async ({ page }) => {
         await page.goto('/admin/settings');
         await expect(page.locator('.loading-page')).toBeHidden();
 
         const bar = page.locator('.settings-titlebar');
-        await expect(bar.locator('.settings-page-title')).toBeVisible();
+        await expect(page.getByRole('heading', { name: '系统设置', exact: true })).toBeVisible();
+        await expect(page.getByRole('tab', { name: '实例设置' })).toHaveAttribute('aria-selected', 'true');
         await expect(bar.getByRole('button', { name: '保存配置' })).toBeVisible();
 
-        // 悬浮固定：向下滚动后标题栏仍停在顶部（贴应用顶栏下方）
+        // 路径页签贴在应用顶栏下方，保存栏继续贴在页签下方。
         await page.evaluate(() => window.scrollTo({ top: 1200 }));
         await expect.poll(async () => {
             return await bar.evaluate(el => Math.round(el.getBoundingClientRect().top));
-        }, { timeout: 3000 }).toBeLessThanOrEqual(56);
+        }, { timeout: 3000 }).toBeLessThanOrEqual(104);
     });
 
     test('左侧分组导航与全部分类渲染', async ({ page }) => {
@@ -34,8 +35,8 @@ test.describe('Admin Settings (GitHub-style layout)', () => {
             await expect(page.locator('.settings-nav-group', { hasText: g })).toBeVisible();
         }
 
-        // 元数据来源和全局设备已迁入插件中心，系统设置保留 15 个分类。
-        await expect(page.locator('.settings-section')).toHaveCount(15);
+        // 元数据来源和全局设备已迁入插件中心，系统设置保留 14 个分类。
+        await expect(page.locator('.settings-section')).toHaveCount(14);
         await expect(page.locator('.settings-nav-item', { hasText: '互联网书籍信息源' })).toHaveCount(0);
         await expect(page.locator('.settings-nav-item', { hasText: '全局设备管理' })).toHaveCount(0);
 
@@ -51,19 +52,14 @@ test.describe('Admin Settings (GitHub-style layout)', () => {
         await expect(page.getByLabel('每本有声书保留的历史版本数')).toHaveValue('3');
     });
 
-    test('基础信息中默认启用并可保存关闭网络书库展示开关', async ({ page }) => {
+    test('旧网络书库展示开关不再隐藏聚合入口', async ({ page }) => {
         await page.goto('/admin/settings');
         await expect(page.locator('.loading-page')).toBeHidden();
 
-        const checkbox = page.getByRole('checkbox', { name: '显示网络书库入口与连载状态筛选' });
-        await expect(checkbox).toBeVisible();
-        await expect(checkbox).toBeChecked();
-
-        await checkbox.uncheck();
-        await page.locator('.settings-titlebar').getByRole('button', { name: '保存配置' }).click();
-        await expect(checkbox).not.toBeChecked();
+        await expect(page.getByRole('checkbox', { name: '显示网络书库入口与连载状态筛选' })).toHaveCount(0);
 
         await page.goto('/');
+        await expect(page.locator('nav').getByRole('link', { name: '书库浏览' })).toBeVisible();
         await expect(page.locator('nav').getByRole('link', { name: '网络书库' })).toHaveCount(0);
     });
 
@@ -76,7 +72,7 @@ test.describe('Admin Settings (GitHub-style layout)', () => {
         // 对应 section 滚动到视口顶部附近
         await expect.poll(async () => {
             return await page.locator('#sec-emailService').evaluate(el => Math.round(el.getBoundingClientRect().top));
-        }, { timeout: 4000 }).toBeLessThan(120);
+        }, { timeout: 4000 }).toBeLessThan(180);
 
         await expect(page.locator('.settings-nav-item.active')).toHaveText('邮件服务');
     });
