+++ app/test/e2e/webdav-readme.spec.ts
@@ -3,30 +3,29 @@ import { test, expect } from '@playwright/test';
 
 const mockApi = process.env.MOCK_API_URL || 'http://127.0.0.1:8080';
 
-test.describe('WebDAV Readme Page', () => {
+test.describe('App access page', () => {
     test.beforeEach(async ({ request }) => {
         await request.post(`${mockApi}/_test/reset`, {
             data: { installed: true }
         });
     });
 
-    test('displays WebDAV connection guide', async ({ page }) => {
-        await page.goto('/webdav-readme');
+    test('combines Moke, OPDS and WebDAV connection details', async ({ page }) => {
+        await page.goto('/library/apps');
 
-        await expect(page.getByRole('heading', { name: 'WebDAV 介绍' })).toBeVisible();
-        await expect(page.getByText('您的 WebDAV 地址为：')).toBeVisible();
-        await expect(page.locator('main').getByText('WebDAV 介绍', { exact: true })).toHaveCount(1);
-
-        // 应包含常见客户端与配置步骤
-        await expect(page.getByText('常见 WebDAV 客户端')).toBeVisible();
-        await expect(page.getByText('配置指南')).toBeVisible();
+        await expect(page.getByRole('heading', { name: '书库浏览' })).toBeVisible();
+        await expect(page.getByRole('tab', { name: '使用 App 访问' })).toHaveAttribute('aria-selected', 'true');
+        await expect(page.getByRole('heading', { name: 'Moke' })).toBeVisible();
+        await expect(page.getByRole('heading', { name: 'OPDS' })).toBeVisible();
+        await expect(page.getByRole('heading', { name: 'WebDAV' })).toBeVisible();
+        await expect(page.getByText(/\/opds\//)).toBeVisible();
+        await expect(page.getByText(/\/books\//)).toBeVisible();
     });
 
-    test('shows WebDAV link on the OPDS readme page', async ({ page }) => {
+    test('redirects both legacy introduction routes to app access', async ({ page }) => {
         await page.goto('/opds-readme');
-
-        const webdavLink = page.getByRole('link', { name: 'WebDAV 介绍', exact: true });
-        await expect(webdavLink).toBeVisible();
-        await expect(webdavLink).toHaveAttribute('href', '/webdav-readme');
+        await expect(page).toHaveURL('/library/apps');
+        await page.goto('/webdav-readme');
+        await expect(page).toHaveURL('/library/apps');
     });
 });
