+++ app/utils/local-mdi-icons.ts
@@ -5,6 +5,7 @@ import {
     mdiAccountCircle,
     mdiAccountCog,
     mdiAccountGroup,
+    mdiAccountOutline,
     mdiAccountRemove,
     mdiAccountVoice,
     mdiAlert,
@@ -38,6 +39,7 @@ import {
     mdiBroom,
     mdiBullhornVariant,
     mdiCellphone,
+    mdiCellphoneLink,
     mdiCheck,
     mdiCheckCircle,
     mdiCheckCircleOutline,
@@ -49,6 +51,7 @@ import {
     mdiClose,
     mdiCloseCircle,
     mdiCloudSearch,
+    mdiCloudSearchOutline,
     mdiCloudSync,
     mdiCloudUpload,
     mdiCodeTags,
@@ -78,6 +81,8 @@ import {
     mdiEarthOff,
     mdiEmail,
     mdiEmailCheck,
+    mdiEmailFastOutline,
+    mdiEmailOutline,
     mdiExitToApp,
     mdiEyeOutline,
     mdiFaceMan,
@@ -93,6 +98,7 @@ import {
     mdiFileSync,
     mdiFileUploadOutline,
     mdiFolder,
+    mdiFolderNetworkOutline,
     mdiFolderOpenOutline,
     mdiFolderOutline,
     mdiFormatListNumbered,
@@ -115,10 +121,12 @@ import {
     mdiLibraryShelves,
     mdiLink,
     mdiLock,
+    mdiLockOutline,
     mdiMagnify,
     mdiMarker,
     mdiMenu,
     mdiMessageAlert,
+    mdiMessageOutline,
     mdiNoteMultipleOutline,
     mdiNoteOutline,
     mdiNoteTextOutline,
@@ -142,10 +150,13 @@ import {
     mdiRssBox,
     mdiScriptTextOutline,
     mdiServer,
+    mdiServerNetwork,
     mdiShieldAccount,
+    mdiShieldAccountOutline,
     mdiShieldCheck,
     mdiShieldCheckOutline,
     mdiShieldEdit,
+    mdiShieldLockOutline,
     mdiShuffle,
     mdiSkipNext,
     mdiSkipPrevious,
@@ -161,6 +172,7 @@ import {
     mdiTextBoxOutline,
     mdiTranslate,
     mdiTrendingUp,
+    mdiTuneVariant,
     mdiUpdate,
     mdiUpload,
     mdiVolumeHigh,
@@ -185,6 +197,7 @@ export const localMdiPaths: Readonly<Record<string, string>> = {
     'mdi-account-circle': mdiAccountCircle,
     'mdi-account-cog': mdiAccountCog,
     'mdi-account-group': mdiAccountGroup,
+    'mdi-account-outline': mdiAccountOutline,
     'mdi-account-remove': mdiAccountRemove,
     'mdi-account-voice': mdiAccountVoice,
     'mdi-alert': mdiAlert,
@@ -218,6 +231,7 @@ export const localMdiPaths: Readonly<Record<string, string>> = {
     'mdi-broom': mdiBroom,
     'mdi-bullhorn-variant': mdiBullhornVariant,
     'mdi-cellphone': mdiCellphone,
+    'mdi-cellphone-link': mdiCellphoneLink,
     'mdi-check': mdiCheck,
     'mdi-check-circle': mdiCheckCircle,
     'mdi-check-circle-outline': mdiCheckCircleOutline,
@@ -229,6 +243,7 @@ export const localMdiPaths: Readonly<Record<string, string>> = {
     'mdi-close': mdiClose,
     'mdi-close-circle': mdiCloseCircle,
     'mdi-cloud-search': mdiCloudSearch,
+    'mdi-cloud-search-outline': mdiCloudSearchOutline,
     'mdi-cloud-sync': mdiCloudSync,
     'mdi-cloud-upload': mdiCloudUpload,
     'mdi-code-tags': mdiCodeTags,
@@ -258,6 +273,8 @@ export const localMdiPaths: Readonly<Record<string, string>> = {
     'mdi-earth-off': mdiEarthOff,
     'mdi-email': mdiEmail,
     'mdi-email-check': mdiEmailCheck,
+    'mdi-email-fast-outline': mdiEmailFastOutline,
+    'mdi-email-outline': mdiEmailOutline,
     'mdi-exit-to-app': mdiExitToApp,
     'mdi-eye-outline': mdiEyeOutline,
     'mdi-face-man': mdiFaceMan,
@@ -273,12 +290,14 @@ export const localMdiPaths: Readonly<Record<string, string>> = {
     'mdi-file-sync': mdiFileSync,
     'mdi-file-upload-outline': mdiFileUploadOutline,
     'mdi-folder': mdiFolder,
+    'mdi-folder-network-outline': mdiFolderNetworkOutline,
     'mdi-folder-open-outline': mdiFolderOpenOutline,
     'mdi-folder-outline': mdiFolderOutline,
     'mdi-format-list-numbered': mdiFormatListNumbered,
     'mdi-format-quote-open': mdiFormatQuoteOpen,
     'mdi-github': mdiGithub,
     'mdi-google': mdiGoogle,
+    'mdi-help-circle-outline': mdiHelpCircleOutline,
     'mdi-history': mdiHistory,
     'mdi-home': mdiHome,
     'mdi-home-group': mdiHomeGroup,
@@ -294,10 +313,12 @@ export const localMdiPaths: Readonly<Record<string, string>> = {
     'mdi-library-shelves': mdiLibraryShelves,
     'mdi-link': mdiLink,
     'mdi-lock': mdiLock,
+    'mdi-lock-outline': mdiLockOutline,
     'mdi-magnify': mdiMagnify,
     'mdi-marker': mdiMarker,
     'mdi-menu': mdiMenu,
     'mdi-message-alert': mdiMessageAlert,
+    'mdi-message-outline': mdiMessageOutline,
     'mdi-note-multiple-outline': mdiNoteMultipleOutline,
     'mdi-note-outline': mdiNoteOutline,
     'mdi-note-text-outline': mdiNoteTextOutline,
@@ -321,10 +342,13 @@ export const localMdiPaths: Readonly<Record<string, string>> = {
     'mdi-rss-box': mdiRssBox,
     'mdi-script-text-outline': mdiScriptTextOutline,
     'mdi-server': mdiServer,
+    'mdi-server-network': mdiServerNetwork,
     'mdi-shield-account': mdiShieldAccount,
+    'mdi-shield-account-outline': mdiShieldAccountOutline,
     'mdi-shield-check': mdiShieldCheck,
     'mdi-shield-check-outline': mdiShieldCheckOutline,
     'mdi-shield-edit': mdiShieldEdit,
+    'mdi-shield-lock-outline': mdiShieldLockOutline,
     'mdi-shuffle': mdiShuffle,
     'mdi-skip-next': mdiSkipNext,
     'mdi-skip-previous': mdiSkipPrevious,
@@ -340,6 +364,7 @@ export const localMdiPaths: Readonly<Record<string, string>> = {
     'mdi-text-box-outline': mdiTextBoxOutline,
     'mdi-translate': mdiTranslate,
     'mdi-trending-up': mdiTrendingUp,
+    'mdi-tune-variant': mdiTuneVariant,
     'mdi-update': mdiUpdate,
     'mdi-upload': mdiUpload,
     'mdi-volume-high': mdiVolumeHigh,
